backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "backend for okhati medical store"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.3.0",
    "black>=25.1.0",
    "email-validator>=2.2.0",
    "fastapi>=0.116.0",
    "httpx>=0.27.2",
    "isort>=6.0.1",
    "nltk>=3.9.1",
    "numpy>=2.0.2",
    "pandas>=2.3.1",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "scikit-learn>=1.7.0",
    "sqlalchemy>=2.0.41",
    "supabase>=2.16.0",
    "tensorflow-cpu>=2.18.0",
    "textblob>=0.19.0",
    "transformers>=4.53.1",
    "uvicorn>=0.35.0",
]


[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
